% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictions.R
\name{add_predictions}
\alias{add_predictions}
\title{Use a batch of models to make predictions on new data}
\usage{
add_predictions(models, new_data, models_to_use = NULL)
}
\arguments{
\item{models}{A list with model objects generated by make_xgb_models.}

\item{new_data}{A dataframe of new cases with predictors as columns. Sample names are row names.}

\item{models_to_use}{Optional vector with subset of names of models to use.}
}
\description{
This function takes a list of models and makes predictions on new data.
}
\examples{
make_new_data_predictions(my_model,"ko_ctnnb1",1,1,my_new_data)
}
\keyword{model}
\keyword{predictions}
