% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_predictions_to_seurat.R
\name{append_predictions_to_seurat}
\alias{append_predictions_to_seurat}
\title{Create a gene expression matrix to predict dependencies}
\usage{
append_predictions_to_seurat(
  scrna_folder,
  scrna_name,
  predictions_table,
  scrna_group_cells_by = "seurat_clusters"
)
}
\arguments{
\item{scrna_folder}{Folder that hosts the Seurat object.}

\item{scrna_name}{Name of the Seurat object (without .rds).}

\item{predictions_table}{The predictions table}

\item{scrna_group_cells_by}{The name of the column in the metadata to group by (default = "seurat_clusters").}
}
\description{
This function creates a gene expression matrix from a list of files, each containing bulk gene expression profiles. It returns a samples-by-genes matrix.
}
\examples{
append_predictions_to_seurat("./data","my_seurat_object")
}
\keyword{predictions}
\keyword{seurat}
