% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{change_colname}
\alias{change_colname}
\title{Changes the value column name of a dataframe}
\usage{
change_colname(df, new_colname)
}
\arguments{
\item{df}{The dataframe.}

\item{new_colname}{The new name.}
}
\description{
This function converts an arxpan ID to a cell line name.
}
\keyword{colname}
