% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{download_data}
\alias{download_data}
\title{Download study files into data folder}
\usage{
download_data(samples_table, work_dir = "./")
}
\arguments{
\item{samples_table}{A samples table with the necessary columns.}
}
\description{
This function reads a samples table and downloads the data
}
\examples{
download_data(samples_table, work_dir)
}
\keyword{data}
\keyword{download}
\keyword{samples}
