% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_demo_samples}
\alias{get_demo_samples}
\title{Get demo samples}
\usage{
get_demo_samples(model, samples = NULL, lineage = NULL, model_data = NULL)
}
\description{
This is a helper function to automatically pick cell lines with opposite predictions
}
