% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Get_Double_Digit_Number}
\alias{Get_Double_Digit_Number}
\title{Vectorized double-digit converter}
\usage{
Get_Double_Digit_Number(x)
}
\description{
This function converts a single digit number to double digits.
}
