% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_legend}
\alias{get_legend}
\title{Extract ggplot legend}
\usage{
get_legend(ggplot_object)
}
\arguments{
\item{ggplot_object}{A ggplot object.}
}
\description{
This function extracts the ggplot legend to plot separately.
}
\examples{
get_legend(my_ggplot)
}
\keyword{ggplot}
\keyword{legend}
