% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_original_data.R
\name{get_original_data}
\alias{get_original_data}
\title{Retrieve the training data}
\usage{
get_original_data(model, data)
}
\arguments{
\item{model}{A model object generated with make_xgb_models}

\item{data}{The dataset that was passed to make_xgb_models}
}
\description{
This function generates the full training data for a given model
}
\examples{
get_original_data(my_models[1],my_data)
}
\keyword{data}
