% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores_table.R
\name{get_scores_table}
\alias{get_scores_table}
\title{Summarize models}
\usage{
get_scores_table(models)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models}
}
\description{
This function generates a table with summary stats of each model
}
\examples{
get_scores_table(my_models)
}
\keyword{scores}
