% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Get a palette of specified size}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The size of the palette.}
}
\description{
This function creates a palette of specified size using equal distant hues.
}
\examples{
gg_color_hue(5)                           
}
\keyword{hues}
