% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{helper_plot_top_contributors}
\alias{helper_plot_top_contributors}
\title{Plot top contributions (internal function)}
\usage{
helper_plot_top_contributors(shap_table, name, n, color = "#52565e")
}
\description{
This function generates a plot of the top contributing features for the top n models.
}
\examples{
plot_top_contributors(shap_table, name, 10)
}
\keyword{plot}
