% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bulk_matrix_from_seurat.R
\name{make_expression_matrix_from_seurat}
\alias{make_expression_matrix_from_seurat}
\title{Create a gene expression matrix to predict dependencies}
\usage{
make_expression_matrix_from_seurat(
  scrna_folder,
  scrna_name,
  scrna_group_cells_by = "seurat_clusters",
  scrna_assay_markers = "SCT",
  scrna_assay_vars = "SCT",
  scrna_test_markers = "MAST",
  scrna_logfc = 0.585,
  scrna_minpct = 0.25,
  scrna_label_prefix = "cluster",
  pseudocount = 0.083,
  qvalue_cutoff = 0.05,
  force_find_markers = FALSE
)
}
\arguments{
\item{scrna_folder}{Folder that hosts the Seurat object.}

\item{scrna_name}{Name of the Seurat object (without .rds).}

\item{scrna_group_cells_by}{The name of the column in the metadata to group by (default = "seurat_clusters").}

\item{scrna_assay_markers}{The assay (e.g. "RNA", "SCT", "alra") used for differential expression analysis.}

\item{scrna_assay_vars}{The assay (e.g. "RNA", "SCT", "alra") used for pre-calculated variable features.}

\item{scrna_test_markers}{The method used (e.g. "MAST", "wilcox", "t") used for differential expression analysis. See help for FindAllMarkers() in Seurat.}

\item{scrna_logfc}{See help for FindAllMarkers() in Seurat.}

\item{scrna_minpct}{See help for FindAllMarkers() in Seurat.}

\item{scrna_label_prefix}{When creating a pseudobulk profile for each subgroup, this is a prefix to add to their labels (default = "cluster").}

\item{pseudocount}{An arbitrary small number to add to gene counts before log transformation. 1 works well for 10X data, 0.083 works well for Drop-seq data (default = 1).}

\item{qvalue_cutoff}{FDR cutoff to consider a marker significant.}

\item{force_find_markers}{By default, finding markers will be skipped if done previously (saved as \link{scrna_name}.\link{scrna_group_cells_by}.markers inside the \link{scrna_folder} folder). If TRUE, finding markers will always happen.}
}
\description{
This function creates a gene expression matrix from a list of files, each containing bulk gene expression profiles. It returns a samples-by-genes matrix.
}
\examples{
show_msg("The number of cells is {cell_count}")
}
\keyword{pseudobulk}
