% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_bulk_matrix.R
\name{make_new_bulk_matrix}
\alias{make_new_bulk_matrix}
\title{Create a gene expression matrix to predict dependencies}
\usage{
make_new_bulk_matrix(bulk_files, pseudocount = 1)
}
\arguments{
\item{bulk_files}{A vector of filenames for gene counts or pre-calculated TPMs.}

\item{pseudocount}{An arbitrary small number to add before log-transformation (default = 1).}
}
\description{
This function creates a gene expression matrix from a list of files, each containing bulk gene expression profiles. It returns a samples-by-genes matrix.
}
\examples{
show_msg("The number of cells is {cell_count}")
}
\keyword{message}
