% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictions.R
\name{make_new_data_predictions}
\alias{make_new_data_predictions}
\title{Use a model to make predictions on new data}
\usage{
make_new_data_predictions(model, name, indx, total, new_data)
}
\arguments{
\item{model}{The model object.}

\item{name}{The name of the perturbation.}

\item{indx}{Integer index used, for progress report.}

\item{total}{Integer of the total number of perturbations passed to this function, for progress report.}

\item{new_data}{A dataframe of new cases with predictors as columns. Sample names are row names.}
}
\description{
This function takes new data, and produces predictions, error estimates, and SHAP values
}
\examples{
make_new_data_predictions(my_model,"ko_ctnnb1",1,1,my_new_data)
}
\keyword{model}
\keyword{predictions}
