% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmap_test.R
\name{mixmap_test}
\alias{mixmap_test}
\title{A test Function}
\usage{
mixmap_test(name = NULL)
}
\arguments{
\item{name}{Your name}
}
\description{
This function tests making a function.
}
\examples{
mixmap_test("John")
}
\keyword{test}
