% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contribution_to_new_samples}
\alias{plot_contribution_to_new_samples}
\title{Plot feature contributions to new sample predictions}
\usage{
plot_contribution_to_new_samples(
  models,
  models_to_use,
  model_data,
  new_samples_to_use = NULL,
  n_features = 5,
  n_columns = 4,
  fixed_axis = TRUE,
  show_error = TRUE,
  highlight_significant = TRUE,
  short_title = TRUE,
  sample_info = NULL
)
}
\arguments{
\item{models}{A list of models generated with make_xgb_models and has appended predictions with add_predictions.}

\item{models_to_use}{A vector of model names to plot.}

\item{model_data}{The training dataset used to generate the model.}

\item{n_features}{The number of top contributors to show their individual contribution. All other predictors will have their contribution combined.}

\item{n_columns}{Number of columns to plot in a grid when plotting more than one sample.}

\item{fixed_axis}{If TRUE, the plot will be set to a fixed scale (-0.05 to 1). Default = FALSE.}

\item{show_error}{If TRUE, a shaded area will be used to visualize prediction interval. Default = TRUE.}

\item{short_title}{If using many columns, set to TRUE to shorten the title of the plot.}

\item{sample_info}{If replace_names is TRUE, this must be supplied.}

\item{replace_names}{If TRUE, the sample name will be replaced using get_cell_line_name. Must supply sample_info.}
}
\description{
This function plots the contribution (Shapley values) of the top features to a sample's predictions, using new samples.
}
\examples{
plot_contribution_to_sample_demo(my_models, c("ko_ctnnb1","ko_myod1"), model_dataset, "soft_tissue")
}
\keyword{contribution}
\keyword{plot}
\keyword{shapley}
