% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_shap_scatter}
\alias{plot_shap_scatter}
\title{Plot Shapley values vs feature values}
\usage{
plot_shap_scatter(
  model,
  name,
  model_data,
  n_features = 4,
  n_columns = 2,
  overlay_predictions = FALSE,
  sample_names = NULL,
  remove_prefix = TRUE,
  sample_colors = NULL,
  sample_info = NULL
)
}
\arguments{
\item{model}{A model generated with make_xgb_models and has appended predictions with add_predictions.}

\item{name}{The name of the perturbation whose prediction we want to plot.}

\item{model_data}{The training dataset used to generate the model.}

\item{n_features}{The number of top contributors to show their individual contribution. All other predictors will have their contribution combined.}

\item{n_columns}{Number of columns to plot in a grid when plotting more than one sample.}

\item{sample_names}{The names of the samples we want to highlight.}

\item{sample_colors}{If highlighting samples (with sample_names), this is a vector of colors to use.}
}
\description{
This function plots a scatter of the contribution (Shapley values) of a feature against the value of that feature for each sample.
}
\examples{
plot_shap_scatter(my_models, model_dataset, "ko_ctnnb1", "my_sample", "red")
}
\keyword{contribution}
\keyword{plot}
\keyword{shapley}
