% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_shap_scatter_for_new_samples}
\alias{plot_shap_scatter_for_new_samples}
\title{Plot SHAP vs Feature Value scatter for new data.}
\usage{
plot_shap_scatter_for_new_samples(
  models,
  models_to_use,
  model_data,
  new_samples_to_use = NULL,
  remove_prefix = TRUE,
  sample_colors = NULL,
  n_features = 6,
  n_columns = 3
)
}
\arguments{
\item{models}{A list of models generated with make_xgb_models and has appended predictions with add_predictions.}

\item{models_to_use}{A vector of model names to plot.}

\item{model_data}{The training dataset used to generate the model.}

\item{sample_colors}{Color vector to use for samples_to_use.}

\item{n_features}{The number of top contributors to show their individual contribution. All other predictors will have their contribution combined.}

\item{n_columns}{Number of columns to plot in a grid when plotting more than one sample.}

\item{samples_to_use}{Optional to highlight specific samples.}

\item{lineage_to_use}{Optional to highlight specific samples of certain lineage.}
}
\description{
This function plots the contribution (Shapley values) against the value of a feature.
}
\examples{
plot_shap_scatter_for_new_samples(my_models, c("ko_ctnnb1","ko_myod1"), model_dataset, samples_to_use = "my_sample")
}
\keyword{contribution}
\keyword{plot}
\keyword{scatter}
\keyword{shapley}
