% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top_contributors}
\alias{plot_top_contributors}
\title{Plot top contributions}
\usage{
plot_top_contributors(
  models,
  models_to_use = NULL,
  data_to_use = "training",
  n_predictors = 10,
  n_columns = 5,
  color = "#52565e"
)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models}

\item{models_to_use}{A vector of model names to restrict the plot to.}

\item{data_to_use}{If "training" (default), it will use the full training data. If anything else, it will use the new predictions (must have used add_predictions before).}

\item{n_predictors}{The maximum number of predictors to show per plot.}

\item{n_columns}{If multiple models are being plotted, how many columns in the plot grid to use.}

\item{color}{The color of the bars in the plot.}
}
\description{
This function generates a plot of the top contributing features for the top n models.
}
\examples{
plot_top_contributors(shap_table, name, 10)
}
\keyword{plot}
