% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerup_observations.R
\name{powerup_process_observations}
\alias{powerup_process_observations}
\title{Process a submitted observations run.}
\usage{
powerup_process_observations(
  cleaned_observations_path,
  schema_path,
  positive_controls_path = NULL,
  negative_controls_path = NULL,
  predictions_path,
  out_observations_dir,
  job_id,
  observation_run_id,
  response_set,
  data_version,
  seed = 1L
)
}
\description{
Process a submitted observations run.
}
