% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_features_lineage.R
\name{prep_features_lineage}
\alias{prep_features_lineage}
\title{Prepare the lineage features table}
\usage{
prep_features_lineage(
  ccle_data,
  our_data = NULL,
  lineage_column_name = "lineage"
)
}
\arguments{
\item{ccle_data}{A data frame with cell line names as rownames and at least one column matching lineage_column_name.}

\item{our_data}{A data frame with cell line names as rownames and at least one column matching lineage_column_name.}

\item{lineage_column_name}{The name of the column that holds the lineage information (default = "lineage").}
}
\description{
This function creates a table with one-hot encoded lineage features
}
\examples{
prep_features_lineage(sample.info)
}
\keyword{features}
\keyword{lineage}
