% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{say_hello}
\alias{say_hello}
\title{Some random new function}
\usage{
say_hello(x)
}
\arguments{
\item{x}{Name of person to say hello to.}
}
\description{
This function prints "hello" to test something.
}
\examples{
say_hello("Mush")
}
\keyword{hello}
