% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{show_msg}
\alias{show_msg}
\title{Display message}
\usage{
show_msg(x)
}
\arguments{
\item{x}{Message to be shown, will be used inside glue() call.}
}
\description{
This function prints a message and calls flush.console() to refresh output
}
\examples{
show_msg("The number of cells is {cell_count}")
}
\keyword{message}
