% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_contributions}
\alias{summarize_contributions}
\title{Summarize SHAP values}
\usage{
summarize_contributions(
  models,
  use_new_data = FALSE,
  min_abs_contribution = 0.05
)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models and with predictions added using add_predictions.}
}
\description{
This function generates one consolidated matrix with SHAP values for each perturbation.
}
\examples{
summarize_contributions(my_models)
}
\keyword{SHAP}
\keyword{contribution}
