% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_error}
\alias{summarize_error}
\title{Summarize prediction error}
\usage{
summarize_error(models)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models and with predictions added using add_predictions.}
}
\description{
This function generates a table with prediction error for each perturbation in a list of models.
}
\examples{
summarize_predictions(my_models)
}
\keyword{error}
\keyword{predictions}
