% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_models}
\alias{summarize_models}
\title{Summarize models}
\usage{
summarize_models(models)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models}
}
\description{
This function generates a table with summary stats of each model
}
\examples{
summarize_models(my_models)
}
\keyword{scores}
