% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_predictions}
\alias{summarize_predictions}
\title{Summarize predictions}
\usage{
summarize_predictions(models)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models and with predictions added using add_predictions.}
}
\description{
This function generates a table with predictions for each perturbation in a list of models.
}
\examples{
summarize_predictions(my_models)
}
\keyword{predictions}
