% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_predictions_in_parallel}
\alias{summarize_predictions_in_parallel}
\title{Summarize predictions in parallel}
\usage{
summarize_predictions_in_parallel(models)
}
\arguments{
\item{models}{A list of model objects generated by make_xgb_models and with predictions added using add_predictions.}
}
\description{
This function generates a table with predictions for each perturbation in a list of models using futures.
}
\examples{
summarize_predictions_in_parallel(my_models)
}
\keyword{predictions}
